+++ tropicube-fallenkingdoms/src/main/java/fr/tropicube/fallenkingdoms/game/HeartState.java
@@ -1,0 +1,3 @@
+package fr.tropicube.fallenkingdoms.game;
+
+public enum HeartState { PROTECTED, VULNERABLE, DESTROYED }