+++ tropicube-fallenkingdoms/src/main/java/fr/tropicube/fallenkingdoms/game/KingdomId.java
@@ -1,0 +1,3 @@
+package fr.tropicube.fallenkingdoms.game;
+
+public enum KingdomId { BLUE, RED, GREEN, YELLOW, ORANGE }